+++ FluentStorage.Tests/Integration/Storage/IStoreTest.DirMisc.cs
@@ -1,0 +1,4 @@
+﻿namespace FluentStorage.Tests.Integration.Storage {
+	public partial class IStoreTest {
+	}
+}